crossplat round 1: foundational crates compile on win-msvc + darwin-arm64
bun_core (win): Fd::uv()/stdin()/stdout()/stderr() read RacyCell<Fd> via .read()
  instead of treating the static as a bare Fd (statics were retyped from
  static mut → RacyCell in the safefn sweep but the cfg(windows) readers
  weren't updated).

bun_sys (mac):
  - Bun__errnoName: SystemErrno::init takes i64 on every POSIX target (the
    'i32 on macOS' comment was stale), keep c_int passthrough only on Windows.
  - socketpair_impl non-Linux branch: Fd::from_native (matches sys.zig
    .fromNative); from_system is Windows-only.
  - c module: drop duplicate sysctlbyname/sockaddr_dl re-exports and the
    macOS UTIME_NOW/OMIT consts (already covered by the BSD-wide re-export
    and the top-level cfg(unix) UTIME_* consts).
  - posix module: drop duplicate BSD-only timeval re-export (covered by the
    cfg(unix) one) and duplicate sysctlbyname wrapper.

bun_uws_sys (win): SocketContext stat_for_digest — bun_windows_sys exposes a
  flat namespace, not windows-rs-style Win32::Storage::FileSystem; add the
  missing WIN32_FILE_ATTRIBUTE_DATA / GET_FILEEX_INFO_LEVELS /
  GetFileAttributesExW bindings to externs.rs and import flat.

bun_uws (win): Socket::fd() — cast LIBUS_SOCKET_DESCRIPTOR (usize on Windows)
  to *mut c_void before Fd::from_system, same as us_socket_t::get_fd().

bun_install (mac): widen S::IWUSR/IWGRP/IWOTH (mode_t = u16 on Darwin) to u32
  before masking st_mode, matching the established pattern in shell/ls.rs.

bun_io (win): make pipe_reader::Loop public (cfg-aliased uws::Loop on POSIX,
  uv::Loop on Windows) so BufferedReaderParent::loop_ implementors in
  bun_runtime can name the trait's actual return type instead of hardcoding
  bun_uws_sys::Loop (== WindowsLoop on Windows ≠ uv::Loop). Updated all 8
  implementors.

bun_runtime: doc-comment-on-fn-param → line comment (rustc rejects /// on
  cfg-gated params once the windows arm is active).

aarch64-apple-darwin: 7 → 0 errors (workspace clean).
x86_64-pc-windows-msvc: bun_core/sys/uws_sys/uws/install/io now compile,
  exposing 416 downstream errors in bun_runtime cfg(windows) paths (never
  reached before) for follow-up rounds. Native linux unchanged (0).

+++ src/install/PackageInstaller.rs
@@ -1982,11 +1982,11 @@ impl<'a> PackageInstaller<'a> {
                                 // `st_mode` is u16 on FreeBSD, u32 elsewhere; widen.
                                 let st_mode = stat.st_mode as u32;
                                 let is_writable = if stat.st_uid == unsafe { bun_sys::c::getuid() } {
-                                    st_mode & bun_sys::S::IWUSR > 0
+                                    st_mode & bun_sys::S::IWUSR as u32 > 0
                                 } else if stat.st_gid == unsafe { bun_sys::c::getgid() } {
-                                    st_mode & bun_sys::S::IWGRP > 0
+                                    st_mode & bun_sys::S::IWGRP as u32 > 0
                                 } else {
-                                    st_mode & bun_sys::S::IWOTH > 0
+                                    st_mode & bun_sys::S::IWOTH as u32 > 0
                                 };
 
                                 if !is_writable {
